Allow only one movement input per frame
This fixes a bug where raplidly changing directions could allow the
player to 180 on themself and die.  This happens because the input
would change the player's velocity before the position was updated.
Since the player was not moved yet, but the velocity has changed, the
check preventing 180 turns would not stop the player.